+++ instagram/src/App.js
@@ -1,32 +1,23 @@
 import React, { Component } from 'react';
 import './App.css';
 
+import Authenticate from './components/Authentication/Authenticate';
 import PostsPage from './components/PostContainer/PostsPage';
-import dummyData from './dummy-data';
 
-export default class App extends Component {
+
+class App extends Component {
   constructor() {
     super();
-    this.state = {
-      posts: [],
-    }
-  }
-
-  componentDidMount() {
-    this.setState({ posts: dummyData });
-  }
-
-  filterPosts = e => {
-    const filteredPosts = dummyData.filter(post =>
-      post.username.includes(e.target.value));
-    this.setState({ posts: filteredPosts });
+    this.state = {};
   }
 
   render() {
     return (
       <div className="app-container">
-        <PostsPage onChange={this.filterPosts} posts={this.state.posts}/>
+        <PostsPage />
       </div>
     );
   }
 }
+
+export default Authenticate(App);
