+++ instagram/src/components/PostContainer/PostsPage.js
@@ -1,15 +1,35 @@
-import React from 'react';
+import React, { Component } from 'react';
+import dummyData from '../../dummy-data';
+
 import SearchBar from '../SearchBar/SearchBar';
 import PostContainer from './PostContainer';
 
-const PostsPage = ({ filterPosts, posts }) => {
+class PostsPage extends Component {
+  constructor() {
+    super();
+    this.state = {
+      posts: [],
+    }
+  }
 
-  return (
-    <div className="app-container">
-      <SearchBar onChange={filterPosts} />
-      <PostContainer posts={posts} />
-    </div>
-  );
+  componentDidMount() {
+    this.setState({ posts: dummyData });
+  }
+
+  filterPosts = e => {
+    const filteredPosts = dummyData.filter(post =>
+      post.username.includes(e.target.value));
+    this.setState({ posts: filteredPosts });
+  }
+
+  render() {
+    return (
+      <div className="app-container">
+        <SearchBar onChange={this.filterPosts} />
+        <PostContainer posts={this.state.posts} />
+      </div>
+    );
+  }
 }
 
 export default PostsPage;
